moves types to '@frontity/wp-source/types'
