+++ packages/wp-source/src/libraries/handlers/date.ts
@@ -1,4 +1,4 @@
-import { Handler } from "../../../";
+import { Handler } from "../../../types";
 import getTotal from "./utils/get-total";
 import getTotalPages from "./utils/get-total-pages";
