+++ packages/wp-source/src/libraries/handlers/post.ts
@@ -1,4 +1,4 @@
-import { Handler } from "../../..";
+import { Handler } from "../../../types";
 import getIdBySlug from "./utils/get-id-by-slug";
 
 const postHandler: Handler = async ({ route, params, state, libraries }) => {
